refactor: remove temp bin to main
