--- src/lib.rs
@@ -1,4 +1,0 @@
-pub mod config;
-pub mod framework;
-pub mod runtime;
-pub mod sketches;
